Add: shoes page, add: custom appbar widget, add: custom widgets, update: main
Add shoes page, this page will be the main page of the application, it returns a scaffold and the CustomAppBar as a body.
Add custom appbar widget, this is the develop of the CustomAppBar.
add custom widgets, this page will contain all the custom widgets to avoid so much importations in the main page (Shoes page).
Update main, in the home of the MaterialApp, add the Shoes Page that as its mentioned before, is the main page of the application.