Add: zapato size, update: custom widgets, update: custom appbar widget, update: shoes page.
Add zapato size, this page will contain the background of the shoe and the shoe over the background, also the shoe will have a shadow, for that reason a stack is used.
Update custom widgets adding the zapato size as a export to use it into the shoes page.
Update custom appbar widget adding parameter called texto that is the text that the user wants to show into the appbar.
Update shoes page adding a Column in the body and the CustomAppBar and ZapatoSizePreview as a children, in this case the text that will be show in the appbar is "For you"