Add: zapato descripcion, update: custom widgets, update: shoes page.
Add zapato descripcion, this widget will contain the information of the shoe, it has 2 required parameters called titulo and descripcion.
Update custom widgets exporting the package of the zapato description.
Update shoes page wrapping the column of the body into a SingleChildScrollView to allow smaller devices to see the information and avoid a oversized in the application.