Add: zapato cart, add: orange button, update: custom widgets, update: shoes page.
Add zapato cart implementing the bottom of the page that consist in a floating button with a text and an orange button, as the orange button will be used in the future in other pages, it will be created as a individual widget, this zapato cart has the monto parameter, this parameter will be show as the text in the row.
Add orange button, this page is the implementation of the orange button, this page will have 4 parameters as texto, buttonWidth, buttonHeight and buttonColor, those parameters works as make the button full customizable.
Update custom widgets, adding the export to zapato cart and orange button.
Update shoes page making a modification in the body from a single child scroll view to a column, this change is with the purpose of keep the ZapatoPriceCart widget always in the bottom of the application no matter if the user does a scroll.