Add: shoes description page, update: zapato size, update: main
Add shoes description page, this page will contain the description of the shoe and the buy button(TODO).
Update zapato size, adding a parameter called fullscreen, for this page this parameter is initialize in false, with this parameter its possible to know when the user is in the cart page or if it is in the buy page, depends on this the page will suffer a few changes, for example the padding will be less, also the height of the container will be shorter, the borderRadius in the top of the image will be different of the bottom and the sizes of the shoes will not be showed.
Update main, implementing the ShoesDescriptionpage in the home of the MaterialApp