Update: pubspec, update: main, update: shoes page, update: shoes description page.
Update pubspec using provider and animate do from the pub.dev.
Update main using the main page ShoesPage as home.
Update shoes page enclosing ZapatoSizePreview in a gesture detector, adding the on tap property to navigate to the ShoesDescriptionPage, also using the Hero widget to make a better looking animation when the navigations is on way.
Update shoes description page enclosing the ZapatoSizePreview in a Hero widget and using the same tag to animate it, also the button over the shoe image implementing to pop the page, the three buttons in the bottom of the screen were animated with a bounce animation, also the buttons of the shoe color options were animated with a fadeinleft using an index  as a property on the duration to get a good animation