add: zapato model, update: main, update: zapato size.
Add zapato model with change notifier to allow the application to use provider with two properties, the talla(shoe size) and the assetImage to change the color of the shoe depends on what color user wants.
Update main enclosing the highest widget of the application in this case MyApp into a MultiProvider with the change notifier provider using the zapato model.
Update zapato size instancing the provider into the build of CajaTallaZapato method taking the talla property to know what shoe size wants the user, for this purpose was neccesary enclose the Padding into a GestureDetecto and in its ontap equaling the talla of the zapato model to the talla of the property into the method.