Update: zapato model, update: shoes description page, update: zapato size
Update zapato model fixing the direction of the assetImage from blue.png to azul.png.
Update shoes description page adding the property urlAsset, this property will save the direction to the assetImage, also instancing provider into the build of the BotonColor method and enclosing the container into a gesture detector to know where the user is tapping and into its ontap method giving the value of the urlAsset to the provider.
Update zapato size implementing the provider and into the AssetImage putting the provider to show the color that the user wants, also turning the ZapatoConSombra to a statefullwidget to precache the images and avoid to show the blink effect.